requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
sentence-transformers
chromadb
python-dotenv
